move window module
